refactor(rate-of-closure): split the execution workspace under the LOC budget
regional_ground_execution_workspace arrives from the source branch at 571
lines, over the repo's 500 LOC budget for changed files. Split along two
existing seams rather than baselined:

  - regional_ground_execution_files_mixin -- the atomic save/export commands
  - regional_ground_execution_status_mixin -- status label and action state

Both stay mixins, matching morris_workspace_mixin and plot_export_mixin,
because every method reads the workspace's own widgets or parents a modal
dialog on it.

The files mixin declares _set_status as an annotation rather than a
NotImplementedError stub: a concrete method there would have shadowed the
status mixin's real implementation through the MRO. Verified at runtime
that the workspace resolves _set_status to the status mixin.

Verified: 3,092 tests pass; 42/42 modules import; file-size budget clean.
